refactor(cron): separate script result parsing
Move existing trigger and payload result parsing into its own module so
runtime ownership changes can be reviewed independently. Preserve parser
bodies, result precedence, state limits, and duration validation.

Related: #140674
